refactor: move cli building into a dedicated internal package
